data-generator/requirements.txt
sseclient==0.0.27
kafka-python==2.0.2
